Manual

Short and clean instructions how to use the Package. Functions and methods can be found here with usage examples.

JuSwalbe.calc_equilibrium_distribution — Function

calc_equilibrium_distribution(mom:macroquant64_1d, gravity=0)

Calculates the equilibrium distributions based on the macroscopic quantities mom and gravitational acceleration gravity.

The equilibrium distribtutions are at the heart of the lattice Boltzmann method. As the expansion is made around the equilibrium, lattice Boltzmann always assumes that the flow field is close to equilibrium. Therefore the equilibrium is calculated from macroscopic quantities, i.e. height .height and velocity .velocity. There are plenty of ways to calculate them I mainly the ones derived in Eq.(13) of Study of the 1D lattice Boltzmann shallow water equation and its coupling to build a canal network.

Math:

Defining equations in one spatial dimensions are expressed in the following way ``\begin{align} f0^{eq} &= h - \frac{1}{2v^2}gh^2 - \frac{1}{v^2}hu^2 \ f1^{eq} &= \frac{1}{4v^2}gh^2 + \frac{1}{2v}hu + \frac{1}{2v^2}hu^2
f_2^{eq} &= \frac{1}{4v^2}gh^2 - \frac{1}{2v}hu + \frac{1}{2v^2}hu^2 \end{align}``

Example

TBD

JuSwalbe.readinput — Method

readinput(file)

Reads input parameters from a file.

The expected amount of parameters can be addressed with inputconstants. For now it expects seven values for different runtime constants.

Example

julia> using JuSwalbe, DelimitedFiles

julia> args = ["Lattice_points_x" 10; "Lattice_points_y" 5; "Max_run_time" 1000; "Output_dump" 100; "gravity" 0.0; "surface_tension" 0.01; "slippage" 1.0] # Generate a text file with input
7×2 Array{Any,2}:
 "Lattice_points_x"    10
 "Lattice_points_y"     5
 "Max_run_time"      1000
 "Output_dump"        100
 "gravity"              0.0
 "surface_tension"      0.01
 "slippage"             1.0

julia> writedlm("test.txt", args)

julia> test = readinput("test.txt")
JuSwalbe.inputconstants(10, 5, 1000, 100, 0.0, 0.01, 1.0)

julia> test.lx
10

julia> test.γ
0.01

julia> test.γ + test.δ
1.01

julia> isa(test.lx + test.gravity, Int32)
false

julia> rm("test.txt")

JuSwalbe.inputconstants — Type

inputconstants = new(lx, ly, maxruntime, dumping, gravity, γ, δ)

Struct containing input parameters.

Contains .lx lattice points in x-direction, .ly lattice points in y-direction. Other fields are .maxruntime for the maximal number of time steps and .dumping to limit the number of output files. On top of these there are physical quantities such as .gravity, .γ and .δ for the values of gravitational acceleration, fluids surface tension and the slip length. The example relates to an quadratic lattice 20 times 20 lattice units in area. Run for 100 lattice Boltzmann time steps only printing output every 10 time steps. Having no gravity and a surface tension of 0.01 and a slip length of 1.

Example

julia> using JuSwalbe

julia> new_input = JuSwalbe.inputconstants(20, 20, 100, 10, 0.0, 0.01, 1.0)
JuSwalbe.inputconstants(20, 20, 100, 10, 0.0, 0.01, 1.0)

julia> new_input.γ
0.01

JuSwalbe.inputfile — Type

inputfile

Abstract type for all kinds of input files