Manual

Short and clean instructions how to use the Package. Functions and methods can be found here with usage examples.

JuSwalbe.calc_equilibrium_distribution — Method

calc_equilibrium_distribution(mom::JuSwalbe.Macroquant; gravity=0)

Calculates the equilibrium distributions based on the macroscopic quantities mom and gravitational acceleration gravity.

The equilibrium distribtutions are at the heart of the lattice Boltzmann method. As the expansion is made around the equilibrium, lattice Boltzmann always assumes that the flow field is close to equilibrium. Therefore the equilibrium is calculated from macroscopic quantities, i.e. height .height and velocity .velocity.

Math:

One spatial dimension

Defining equations in one spatial dimensions are expressed in the following way

$f_0^{eq} = h - \frac{1}{2v^2}gh^2 - \frac{1}{v^2}hu^2$

$f_1^{eq} = \frac{1}{4v^2}gh^2 + \frac{1}{2v}hu + \frac{1}{2v^2}hu^2$

$f_2^{eq} = \frac{1}{4v^2}gh^2 - \frac{1}{2v}hu + \frac{1}{2v^2}hu^2$

Two spatial dimension

In two spatial dimensions the velocity because a vector with a x and y component. Such the equations look a little more complex

$f_0^{eq} = h - \frac{4}{9}h(\frac{15}{2}gh - \frac{3}{2}u^2)$

$f_i^{eq} = w_i h(\frac{3}{2}gh + 3 \mathbf{c}_i\cdot\mathbf{u} + \frac{9}{2}(\mathbf{c}_i\cdot\mathbf{u})^2 - \frac{3}{2}u^2))$

with i are the number of lattice speeds and wi and ci are the weights and sets of lattice velocities.

Example

TBD

References

One spatial dimension

There are plenty of ways to calculate them I mainly the ones derived in Eq.(13) of Study of the 1D lattice Boltzmann shallow water equation and its coupling to build a canal network.

Two spatial dimensions

Here I stick to the paper written by Paul Dellar, Eq.(26) of Nonhydrodynamic modes and a priori construction of shallow water lattice Boltzmann equations Originally these equilibria have been worked out by Paul Salmon.

JuSwalbe.pressure — Method

pressure(mom::JuSwalbe.Macroquant; γ = 0.01, θ = 1/9)

Film pressure of the thin film equation.

This is just the summation of the laplacian of the height Δh and the disjoining potential Π. Since this function uses Π one has to be careful of θ as it should be an array. However it can be either a single element array or as large as the whole domain. See also: Δh, Π

Math

The film pressure is simply

$p_{film} = -\gamma\Delta h + \Pi(h)$

The sign is not 100% fixed and can change from paper to paper. Here I choose to be in agreement with Thiele et al., which is known to be a good theoretician.

Example

julia> using JuSwalbe

julia> height = reshape([i for i in 1.0:1.0:16.0],4,4)
4×4 Array{Float64,2}:
 1.0  5.0   9.0  13.0
 2.0  6.0  10.0  14.0
 3.0  7.0  11.0  15.0
 4.0  8.0  12.0  16.0

julia> p = pressure(height)
4×4 Array{Float64,2}:
 -0.200016  -0.0400001   -0.04        0.12
 -0.160002  -7.44536e-8  -1.6082e-8   0.16
 -0.160001  -4.68862e-8  -1.20826e-8  0.16
 -0.12       0.04         0.04        0.2

So what does this mean? A positive pressure is force that drives the film down in height. While a negative pressure generates a flux towards that location. In the example a linear increasing height field was used, an equilibrium though is reached when the film is flat.

References

Recent (short)

Signatures of slip in dewetting polymer films

Review

Dynamics and stability of thin liquid films

JuSwalbe.readinput — Method

readinput(file)

Reads input parameters from a file.

The expected amount of parameters can be addressed with inputconstants. For now it expects seven values for different runtime constants.

Example

julia> using JuSwalbe, DelimitedFiles

julia> args = ["Lattice_points_x" 10; "Lattice_points_y" 5; "Max_run_time" 1000; "Output_dump" 100; "gravity" 0.0; "surface_tension" 0.01; "slippage" 1.0] # Generate a text file with input
7×2 Array{Any,2}:
 "Lattice_points_x"    10
 "Lattice_points_y"     5
 "Max_run_time"      1000
 "Output_dump"        100
 "gravity"              0.0
 "surface_tension"      0.01
 "slippage"             1.0

julia> writedlm("test.txt", args)

julia> test = readinput("test.txt")
JuSwalbe.inputconstants(10, 5, 1000, 100, 0.0, 0.01, 1.0)

julia> test.lx
10

julia> test.γ
0.01

julia> test.γ + test.δ
1.01

julia> isa(test.lx + test.gravity, Int32)
false

julia> rm("test.txt")

JuSwalbe.Δh — Method

Δh(mom::JuSwalbe.Macroquant)

Calculates the laplacian of the height field with periodic boundaries.

The calculation of the laplacian is central for the thin film evolution. Only the pressure gradient will induce a flow, at least for the non-fluctuating version. Therefore it is fairly important to have an accurate computation of the laplacian.

Math

The Laplace equation is given by

$\Delta \rho = 0$

The laplace operator is simply Δ = ∇ . ∇. Since the film pressure has a contribution Δh we need a discreticed laplace operator.

Two spatial dimensions

In two dimensions I follow the paper from Santosh and Succi. A nine-point stencil is used and the equation goes as follow

$\Delta h_{i,j} = \frac{1}{6}[4\sum_{nn}h_{i,j} + sum_{diag}h_{i,j} - 20 h_{i,j}]$

Where i and j are the x and y coordinates. The index nn relates to the nearest neighbors, all four elements which are exactly Δx away from (i,j). On the other hand the diagonal elements are those four which are have a distance √2Δx to (i,j). Periodicity is taken care of by a circular padding of the height array.

One spatial dimension

For a single spatial dimension it is mostly sufficient to use the central difference approach. This approach is given by

`` \partialx^2 h{i} = h{i-1} - 2h{i} + h_{i+1}

with i being the index along the spatial dimension.

Example

julia> using JuSwalbe

julia> height = reshape(collect(1.0:16.0), (4,4))
4×4 Array{Float64,2}:
 1.0  5.0   9.0  13.0
 2.0  6.0  10.0  14.0
 3.0  7.0  11.0  15.0
 4.0  8.0  12.0  16.0

julia> velx = zeros(Float64, (4,4))
4×4 Array{Float64,2}:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

julia> vely = zeros(Float64, (4,4))
4×4 Array{Float64,2}:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 
julia> vel = JuSwalbe.Twovector(velx, vely)
JuSwalbe.Twovector{Array{Float64,2}}([0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0], [0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0])

julia> moment = JuSwalbe.Macroquant(height, vel, zeros(4,4), zeros(4,4))
JuSwalbe.Macroquant{Array{Float64,2},JuSwalbe.Twovector{Array{Float64,2}}}([1.0 5.0 9.0 13.0; 2.0 6.0 10.0 14.0; 3.0 7.0 11.0 15.0; 4.0 8.0 12.0 16.0], JuSwalbe.Twovector{Array{Float64,2}}([0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0], [0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0]), [0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0], [0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0])

julia> moment.height
4×4 Array{Float64,2}:
 1.0  5.0   9.0  13.0
 2.0  6.0  10.0  14.0
 3.0  7.0  11.0  15.0
 4.0  8.0  12.0  16.0

julia> laplace = Δh(moment)
4×4 Array{Float64,2}:
 20.0   4.0   4.0  -12.0
 16.0   0.0   0.0  -16.0
 16.0   0.0   0.0  -16.0
 12.0  -4.0  -4.0  -20.0

julia> height = collect(1.0:16.0)
16-element Array{Float64,1}:
  1.0
  2.0
  3.0
  4.0
  5.0
  6.0
  7.0
  8.0
  9.0
 10.0
 11.0
 12.0
 13.0
 14.0
 15.0
 16.0

julia> Δh(height)
16-element Array{Float64,1}:
  16.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
   0.0
 -16.0

References

Two spatial dimensions

There are plenty of papers concerning discret differentail operators. Many of them are good, although definitly not the best I go here with:

Isotropic discrete Laplacian operators from lattice hydrodynamics

One spatial dimension

Almost any reference on discrete differentiation is good enough.

JuSwalbe.Π — Method

Π(mom::JuSwalbe.Macroquant; h_star = 0.1, exponents = [9,3], γ = 0.01, θ = 1.0/9.0)

Calculates the disjoining pressure for a given surface tension and contact angle at every lattice point.

The disjoing pressure potential enables the simulation to study phenomena like dewetting of thin liquid films. Only with this potential it is possible to "dewett" without using highly sophisticated boundary conditions at $h\rightarrow 0$. The functional shape of this potential can vary, ranging from the powerlaw used here to exponentials to a mixture of both. Using the powerlaw shape the exponents (9,3) mimic the Lenard-Jones potential for the liquid substrate interaction.

Arguments

mom::JuSwalbe.Macroquant : Macroscopic moments of the simulation, important here .height which contains the height field. Allows for array input as well.

h_star::T : Minimum of the pontential. T needs to be a subtype of Number!

exponents::Array{Int64,1} : Exponents for the powerlaw potential

γ::T : Surface tension

θ::T : Equilibrium contact angle in multiple of π (for accuarcy reasons). Needs to be an array when doing patterning. For simple substrate use a one entry array with correct dimension.

Math

The potential can be derived from the assumption that a given surface energy demands an equilibrium contact angle for the fluid. For the exact derivation take a look at the references, in principle it is the derivative of the interfacial potential with respect to h.

$\Phi'(h) = \Pi(h)$

$\Pi(h) = \kappa f(h) = (1-\cos(\theta))\frac{(n-1)(m-1)}{(n-m)h_{\ast}}\Bigg[\Bigg( rac{h_{\ast}}{h}\Bigg)^n - \Bigg( rac{h_{\ast}}{h}\Bigg)^m\Bigg]$

$\Pi(h_{\ast}) = 0$

Example

julia> using JuSwalbe

julia> n,m = (4,4)
(4, 4)

julia> moment = JuSwalbe.Macroquant{Matrix{Float64}, JuSwalbe.Twovector{Matrix{Float64}}}(ones(n,m), JuSwalbe.Twovector(zeros(n,m),zeros(n,m)), zeros(n,m),zeros(n,m))
JuSwalbe.Macroquant{Array{Float64,2},JuSwalbe.Twovector{Array{Float64,2}}}([1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0], JuSwalbe.Twovector{Array{Float64,2}}([0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0], [0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0]), [0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0], [0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 0.0])

julia> p = Π(moment)
4×4 Array{Float64,2}:
 -1.6082e-5  -1.6082e-5  -1.6082e-5  -1.6082e-5
 -1.6082e-5  -1.6082e-5  -1.6082e-5  -1.6082e-5
 -1.6082e-5  -1.6082e-5  -1.6082e-5  -1.6082e-5
 -1.6082e-5  -1.6082e-5  -1.6082e-5  -1.6082e-5

julia> p = Π(moment, θ=zeros(1,1)) # Fully wetting substrate
4×4 Array{Float64,2}:
 -0.0  -0.0  -0.0  -0.0
 -0.0  -0.0  -0.0  -0.0
 -0.0  -0.0  -0.0  -0.0
 -0.0  -0.0  -0.0  -0.0

julia> h = reshape([i for i in 1.0:1.0:(n*m)],n,m)
4×4 Array{Float64,2}:
 1.0  5.0   9.0  13.0
 2.0  6.0  10.0  14.0
 3.0  7.0  11.0  15.0
 4.0  8.0  12.0  16.0

julia> p = Π(h)
4×4 Array{Float64,2}:
 -1.6082e-5   -1.28656e-7  -2.20603e-8  -7.31997e-9
 -2.01025e-6  -7.44536e-8  -1.6082e-8   -5.86078e-9
 -5.95628e-7  -4.68862e-8  -1.20826e-8  -4.76503e-9
 -2.51281e-7  -3.14101e-8  -9.30669e-9  -3.92626e-9

References

To get a good understanding:

Long-scale evolution of thin liquid films

Wetting and spreading

Dynamics and stability of thin liquid films

A rather recent new setup for the shape of Π can be found in

Signatures of slip in dewetting polymer films

JuSwalbe.velocitysquared — Method

velocitysquared(mom::Macroquant{Matrix{T}, JuSwalbe.Twovector{Matrix{T}}})

Computes the square of the velocity vector (ux, uy) at every lattice point.

The magnitude of the velocity is needed to calculate the equilibrium distribution in the dimensional case. In the one dimensional case the velocity is just a vector and therefore has no .x and .y component. See also calc_equilibrium_distribution

Math

The velocity squared $u^2(x,y)$ is computed according to

$u^2(x,y) = (u_x, u_y)^2(x,y) = u_x^2(x,y) + u_y^2(x,y).$

With lower case x and y the respective component of the velocity vector is addressed.

Example

julia> using JuSwalbe

julia> velocities = JuSwalbe.Twovector{Matrix{Float64}}(fill(0.1, (4,4)),fill(0.2, (4,4)))
JuSwalbe.Twovector{Array{Float64,2}}([0.1 0.1 0.1 0.1; 0.1 0.1 0.1 0.1; 0.1 0.1 0.1 0.1; 0.1 0.1 0.1 0.1], [0.2 0.2 0.2 0.2; 0.2 0.2 0.2 0.2; 0.2 0.2 0.2 0.2; 0.2 0.2 0.2 0.2])

julia> moment = JuSwalbe.Macroquant{Matrix{Float64},JuSwalbe.Twovector{Matrix{Float64}}}(ones(Float64, (4,4)), velocities, ones(Float64, (4,4)), ones(Float64, (4,4)))
JuSwalbe.Macroquant{Array{Float64,2},JuSwalbe.Twovector{Array{Float64,2}}}([1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0], JuSwalbe.Twovector{Array{Float64,2}}([0.1 0.1 0.1 0.1; 0.1 0.1 0.1 0.1; 0.1 0.1 0.1 0.1; 0.1 0.1 0.1 0.1], [0.2 0.2 0.2 0.2; 0.2 0.2 0.2 0.2; 0.2 0.2 0.2 0.2; 0.2 0.2 0.2 0.2]), [1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0], [1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0; 1.0 1.0 1.0 1.0])

julia> moment.velocity.x
4×4 Array{Float64,2}:
 0.1  0.1  0.1  0.1
 0.1  0.1  0.1  0.1
 0.1  0.1  0.1  0.1
 0.1  0.1  0.1  0.1

julia> JuSwalbe.velocitysquared(moment)
4×4 Array{Float64,2}:
 0.05  0.05  0.05  0.05
 0.05  0.05  0.05  0.05
 0.05  0.05  0.05  0.05
 0.05  0.05  0.05  0.05

JuSwalbe.inputconstants — Type

inputconstants = new(lx, ly, maxruntime, dumping, gravity, γ, δ)

Struct containing input parameters.

Contains .lx lattice points in x-direction, .ly lattice points in y-direction. Other fields are .maxruntime for the maximal number of time steps and .dumping to limit the number of output files. On top of these there are physical quantities such as .gravity, .γ and .δ for the values of gravitational acceleration, fluids surface tension and the slip length. The example relates to an quadratic lattice 20 times 20 lattice units in area. Run for 100 lattice Boltzmann time steps only printing output every 10 time steps. Having no gravity and a surface tension of 0.01 and a slip length of 1.

Example

julia> using JuSwalbe

julia> new_input = JuSwalbe.inputconstants(20, 20, 100, 10, 0.0, 0.01, 1.0)
JuSwalbe.inputconstants(20, 20, 100, 10, 0.0, 0.01, 1.0)

julia> new_input.γ
0.01

JuSwalbe.inputfile — Type

inputfile

Abstract type for all kinds of input files